fix bug stack linkedlist, rename package uppercase --> lower case
