+++ src/main/java/doublelinkedlist/DoubleLinkedList.java
@@ -1,4 +1,4 @@
-package DoubleLinkedList;
+package doublelinkedlist;
 
 public class DoubleLinkedList {
     Node head;
