+++ src/main/java/doublelinkedlist/Node.java
@@ -1,4 +1,4 @@
-package DoubleLinkedList;
+package doublelinkedlist;
 
 public class Node {
     int data;
