+++ src/main/java/linkedlist/Node.java
@@ -1,4 +1,4 @@
-package Linkedlist;
+package linkedlist;
 
 public class Node {
     int data;
@@ -15,5 +15,13 @@ public class Node {
 
     public Node getNext() {
         return next;
+    }
+
+    public void setNext(Node next) {
+        this.next = next;
+    }
+
+    public void setData(int data) {
+        this.data = data;
     }
 }
