+++ src/main/java/linkedlist/SingleLinkedList.java
@@ -1,4 +1,4 @@
-package Linkedlist;
+package linkedlist;
 
 public class SingleLinkedList {
     Node head;
