+++ src/main/java/stack/StackLinkedList.java
@@ -1,4 +1,4 @@
-package Stack;
+package stack;
 
 
 public class StackLinkedList {
