+++ src/test/java/doublelinkedlist/DoubleLinkedListTest.java
@@ -1,4 +1,4 @@
-package DoubleLinkedList;
+package doublelinkedlist;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
