+++ src/test/java/linkedlist/SingleLinkedListTest.java
@@ -1,13 +1,10 @@
-package LinkedList;
+package linkedlist;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
-import Linkedlist.Node;
-import Linkedlist.SingleLinkedList;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
 public class SingleLinkedListTest {
