+++ src/test/java/stack/StackArrayListTest.java
@@ -1,4 +1,4 @@
-package Stack;
+package stack;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -6,35 +6,37 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class StackLinkedListTest {
-    StackLinkedList stackLinkedList;
+public class StackArrayListTest
+{
+    StackArrayList stackArrayList;
     @BeforeEach
     void setUp() {
-        stackLinkedList = new StackLinkedList(1);
+        stackArrayList = new StackArrayList(1);
     }
 
     @Test
     @DisplayName("Test top method of stack array list")
     void testTop() {
-        assertEquals(1, stackLinkedList.top(),
+        assertEquals(1, stackArrayList.top(),
                 "Get element method is incorrect");
     }
 
     @Test
     @DisplayName("Test push method of stack array list")
     void testPush() {
-        stackLinkedList.push(2);
-        assertEquals(2, stackLinkedList.top(),
+        stackArrayList.push(2);
+        assertEquals(2, stackArrayList.top(),
                 "Get element method is incorrect");
     }
 
     @Test
     @DisplayName("Test pop method of stack array list")
     void testPop() {
-        stackLinkedList.push(2);
-        stackLinkedList.push(3);
-        stackLinkedList.pop();
-        assertEquals(2, stackLinkedList.getSize(),
+        stackArrayList.push(2);
+        stackArrayList.push(3);
+        stackArrayList.pop();
+        assertEquals(2, stackArrayList.getSize(),
                 "Get element method is incorrect");
     }
+
 }
